data/requirements.txt
python-slugify==4.0.1